Moved decorate in to util. Cleanup from some PR feedback
