+++ libs/blocks/media/media.css
@@ -126,11 +126,13 @@
   padding: var(--spacing-s) 0;
   position: relative;
 }
+
 .media[class*="-up"] .foreground > .media-row {
   width: 100%;
   display: block;
   max-width: 100%;
 }
+
 .media[class*="-up"] .foreground > .media-row > div:first-of-type {
   margin-right: 0;
   margin-bottom: var(--spacing-s);
@@ -206,13 +208,16 @@
   .media > .foreground .media-row {
     gap: var(--spacing-l);
   }
+
   .media.small > .foreground .media-row {
     gap: var(--spacing-m);
   }
+
   .media.large > .foreground .media-row {
     gap: var(--spacing-xl);
     margin: 0 auto;
   }
+
   .media.large > .foreground .media-row .image { min-width: 700px; }
   .media[class*="-up"].large > .foreground .media-row .image { min-width: 100%; }
   .media.three-up .foreground { grid-template-columns: repeat(3, 1fr); }
