+++ libs/blocks/media/media.js
@@ -14,15 +14,13 @@
  * Media - v5.1
  */
 
-import { decorateBlockBg, decorateBlockDaa, decorateText, getBlockSize } from '../../scripts/decorate.js';
+import { decorateBlockBg, decorateBlockDaa, decorateText, getBlockSize } from '../../utils/utils.js';
 
 export default function init(el, daa = true) {
   if (daa) decorateBlockDaa(el);
   const children = el.querySelectorAll(':scope > div');
-  if (children.length > 1) {
-    if (children[0].childElementCount === 1) {
-      decorateBlockBg(el, children[0]);
-    }
+  if (children.length && children[0].childElementCount === 1) {
+    decorateBlockBg(el, children[0]);
   }
   const size = getBlockSize(el);
   const media = el.querySelectorAll(':scope > div:not([class])');
@@ -39,7 +37,7 @@ export default function init(el, daa = true) {
     const image = row.querySelector(':scope > div:not([class])');
     if (image) image.classList.add('image');
     const img = image.querySelector(':scope img');
-    if (header && img && img.alt === '') img.alt = header.textContent;
+    if (header && img?.alt === '') img.alt = header.textContent;
     container.append(row);
   });
   el.append(container);
